requirements.txt
aiofiles==23.2.1
aiogram==3.7.0
aiohttp==3.9.5
aiosignal==1.3.1
aiosqlite==0.20.0
annotated-types==0.7.0
async-timeout==4.0.3
attrs==23.2.0
beautifulsoup4==4.12.3
bs4==0.0.2
cachetools==5.3.3
certifi==2024.6.2
charset-normalizer==3.3.2
frozenlist==1.4.1
google-auth==2.30.0
google-auth-oauthlib==1.2.0
greenlet==3.0.3
gspread==6.1.2
idna==3.7
magic-filter==1.0.12
multidict==6.0.5
oauthlib==3.2.2
pyasn1==0.6.0
pyasn1_modules==0.4.0
pydantic==2.7.3
pydantic_core==2.18.4
python-dotenv==1.0.1
requests==2.32.3
requests-oauthlib==2.0.0
rsa==4.9
schedule==1.2.2
soupsieve==2.5
SQLAlchemy==2.0.30
typing_extensions==4.12.1
urllib3==2.2.1
yarl==1.9.4
